+++ src/ForceFeedback.Core/Feedbacks/InsertTextFeedback.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace ForceFeedback.Core
+namespace ForceFeedback.Core.Feedbacks
 {
     public class InsertTextFeedback : IFeedback
     {
