+++ src/ForceFeedback.Core/ForceFeedbackMachine.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ForceFeedback.Core.Feedbacks;
+using System;
 using System.Collections.Generic;
 using System.Drawing;
